Enable edgedb-js in browsers without native BigInt
It's now possible to integrate the JSBI library to make edgedb-js
run in browsers without BigInt (notably Safari).

The integration is compatible with webpack & nodejs and can be
enabled via:

   import JSBI from "jsbi";

   if (typeof BigInt === "undefined") {
       plugJSBI(JSBI);
   }

+++ src/codecs/datetime.ts
@@ -141,6 +141,6 @@ export class DurationCodec extends ScalarCodec implements ICodec {
     if (months !== 0) {
       throw new Error("non-zero reserved bytes in duration");
     }
-    return Duration.fromMicroseconds(us);
+    return Duration.fromMicroseconds(us as bigint);
   }
 }
